refactor(config): split load_and_lock pure helpers from ConfigCore
- Extract constants, yaml_io, overrides, digest, and validate modules
- Keep ConfigCore as orchestration-only load_and_lock façade
- Point callers/tests at new pure-helper import paths (Refs: #31)
